+++ CBRN_Project/CBRN_Project/MVVM/Models/Input/Uniform.cs
@@ -8,7 +8,7 @@ namespace CBRN_Project.MVVM.Models
 {
     public class Uniform
     {
-        private float thresoldThermalFluence { get; set; }
+        private double thresoldThermalFluence { get; set; }
 
         private string uniformType { get; set; }
         public string UniformType
